# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StreetClaim RCM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /encounters:
    get:
      operationId: listEncounters
      tags: [encounters]
      summary: List encounters
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending_review, submitted, denied, paid]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of encounters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncounterListResponse"
    post:
      operationId: createEncounter
      tags: [encounters]
      summary: Create a new field encounter (defaults POS 27)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEncounterRequest"
      responses:
        "201":
          description: Encounter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Encounter"
...
  /eligibility/aevs:
    post:
      operationId: aevsLookup
      tags: [eligibility]
      summary: Look up Medi-Cal eligibility via AEVS (X12 270/271)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AevsLookupRequest"
      responses:
        "200":
          description: Eligibility found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AevsLookupResponse"
        "422":
          description: Patient not found or invalid lookup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Encounter:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        providerId:
          type: integer
        encounterDate:
          type: string
          format: date-time
        posCode:
          type: string
          default: "27"
        emCode:
          type: string
          enum: [99211, 99212, 99213, 99214, 99215]
        diagnosisCodes:
          type: array
          items:
            type: string
        sdohZCodes:
          type: array
          items:
            type: string
        clinicalNotes:
          type: string
        locationDescription:
          type: string
        status:
          type: string
          enum: [draft, pending_review, submitted, denied, paid]
        isOfflineCapture:
          type: boolean
        offlineTimestamp:
          type: string
          format: date-time
        syncTimestamp:
          type: string
          format: date-time
        cryptoHash:
          type: string
        payerType:
          type: string
          enum: [medi_cal, medicare, managed_care, private, uninsured]
        reimbursementRate:
          type: string
          enum: [facility, non_facility]
        estimatedReimbursement:
          type: number
        claimId:
          type: integer
        athenaEncounterId:
          type: string
          description: Athenahealth encounter ID after successful EHR sync
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - encounterDate
        - posCode
        - emCode
        - diagnosisCodes
        - sdohZCodes
        - status
        - isOfflineCapture
        - payerType
        - reimbursementRate
        - createdAt
        - updatedAt
    CreateEncounterRequest:
      type: object
      properties:
        patientId:
          type: integer
        encounterDate:
          type: string
          format: date-time
        posCode:
          type: string
          default: "27"
        emCode:
          type: string
          enum: [99211, 99212, 99213, 99214, 99215]
        diagnosisCodes:
          type: array
          items:
            type: string
        clinicalNotes:
          type: string
        locationDescription:
          type: string
        isOfflineCapture:
          type: boolean
          default: false
        offlineTimestamp:
          type: string
          format: date-time
        payerType:
          type: string
          enum: [medi_cal, medicare, managed_care, private, uninsured]
      required:
        - patientId
        - encounterDate
        - emCode
        - payerType
...
    EncounterListResponse:
      type: object
      properties:
        encounters:
          type: array
          items:
            $ref: "#/components/schemas/Encounter"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [encounters, total, page, limit]

    # ─── CLAIM SCHEMAS ─────────────────────────────────────────────────────────
...
    AevsLookupRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
      required: [firstName, lastName, dateOfBirth]
    AevsLookupResponse:
      type: object
      properties:
        cin:
          type: string
          description: Medi-Cal Client Index Number
        managedCarePlan:
          type: string
          description: Managed Care Plan name
        eligibilityStatus:
          type: string
          enum: [active, inactive, pending, unknown]
        lookupDate:
          type: string
          format: date-time
        isStub:
          type: boolean
          description: True when returned from development stub (no live AEVS credentials)
      required: [cin, eligibilityStatus, lookupDate]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]

    # ─── ATHENA SCHEMAS ─────────────────────────────────────────────────────────
...